splitting color functions to sub modules for better separation
