+++ pwndbg/hexdump.py
@@ -9,7 +9,7 @@ from __future__ import unicode_literals
 import copy
 import string
 
-import pwndbg.color
+import pwndbg.color.hexdump as H
 import pwndbg.config
 
 
@@ -30,20 +30,20 @@ def load_color_scheme():
     # We want to colorize the hex characters and only print out
     # printable values on the righ hand side.
     #
-    color_scheme = {i:pwndbg.color.hexdump_normal("%02x" % i) for i in range(256)}
-    printable = {i:pwndbg.color.hexdump_normal('.') for i in range(256)}
+    color_scheme = {i:H.normal("%02x" % i) for i in range(256)}
+    printable = {i:H.normal('.') for i in range(256)}
 
     for c in bytearray((string.ascii_letters + string.digits + string.punctuation).encode('utf-8', 'ignore')):
-        color_scheme[c] = pwndbg.color.hexdump_printable("%02x" % c)
-        printable[c] = pwndbg.color.hexdump_printable("%s" % chr(c))
+        color_scheme[c] = H.printable("%02x" % c)
+        printable[c] = H.printable("%s" % chr(c))
 
     for c in bytearray(b'\x00'):
-        color_scheme[c] = pwndbg.color.hexdump_zero("%02x" % c)
-        printable[c] = pwndbg.color.hexdump_zero('.')
+        color_scheme[c] = H.zero("%02x" % c)
+        printable[c] = H.zero('.')
 
     for c in bytearray(b'\xff\x7f\x80'):
-        color_scheme[c] = pwndbg.color.hexdump_special("%02x" % c)
-        printable[c] = pwndbg.color.hexdump_special('.')
+        color_scheme[c] = H.special("%02x" % c)
+        printable[c] = H.special('.')
 
     color_scheme[-1] = '  '
     printable[-1] = ' '
@@ -68,9 +68,9 @@ def hexdump(data, address = 0, width = 16, skip = True):
         hexline = []
 
         if address:
-            hexline.append(pwndbg.color.hexdump_offset("+%04x " % (i*width)))
+            hexline.append(H.offset("+%04x " % (i*width)))
 
-        hexline.append(pwndbg.color.hexdump_offset("%#08x  " % (base + (i*width))))
+        hexline.append(H.offset("%#08x  " % (base + (i*width))))
 
         for group in groupby(line, 4):
             for char in group:
@@ -78,11 +78,11 @@ def hexdump(data, address = 0, width = 16, skip = True):
                 hexline.append(' ')
             hexline.append(' ')
 
-        hexline.append(pwndbg.color.hexdump_separator('|'))
+        hexline.append(H.separator('|'))
         for group in groupby(line, 4):
             for char in group:
                 hexline.append(printable[char])
-            hexline.append(pwndbg.color.hexdump_separator('|'))
+            hexline.append(H.separator('|'))
 
 
         yield(''.join(hexline))
@@ -90,8 +90,8 @@ def hexdump(data, address = 0, width = 16, skip = True):
     hexline = []
 
     if address:
-        hexline.append(pwndbg.color.hexdump_offset("+%04x " % len(data)))
+        hexline.append(H.offset("+%04x " % len(data)))
 
-    hexline.append(pwndbg.color.hexdump_offset("%#08x  " % (base + len(data))))
+    hexline.append(H.offset("%#08x  " % (base + len(data))))
 
     yield ''.join(hexline)
