+++ pwndbg/ui.py
@@ -12,7 +12,7 @@ import sys
 import termios
 
 import pwndbg.arch
-import pwndbg.color
+import pwndbg.color.context as C
 
 
 def banner(title):
@@ -22,7 +22,7 @@ def banner(title):
     except:
         width = 80
     width -= 2
-    return pwndbg.color.banner(("[{:-^%ss}]" % width).format(title))
+    return C.banner(("[{:-^%ss}]" % width).format(title))
 
 def addrsz(address):
     address = int(address) & pwndbg.arch.ptrmask
